+++ src/res/ext/bigbrain.json
@@ -659,6 +659,12 @@
       "@id": "juelich/iav/atlas/v1.0.0/3",
       "name": "Cortical Layers Segmentation",
       "ngId": "cortical layers",
+      "originDatasets": [
+        {
+          "kgSchema": "minds/core/dataset/v1.0.0",
+          "kgId": "13e6ec48-4fec-4e9a-9bcf-45036d3c8ef9"
+        }
+      ],
       "properties": {
         "name": "Cortical Layers Segmentation",
         "description": "The cerebral isocortex has six cytoarchitectonic layers that vary depending on cortical area and local morphology. This datasets provides a 3D segmentation of all cortical and laminar surfaces in the BigBrain, a high-resolution, 3D histological model of the human brain. The segmentation has been computed automatically based on histological intensities along 3D cortical profiles which were sampled between the pial and white matter throughout the dataset. These cortical profiles were segmented into layers using a convolutional neural network. Training profiles were generated from examples of manually segmented layers on cortical regions from 2D histological sections of the BigBrain. From the segmented intensity profiles, surface meshes and voxel masks of all six cortical layers in the space of the Big Brain have been computed.",
