+++ src/ui/databrowserModule/preview/previewCard/previewCard.template.html
@@ -4,11 +4,41 @@
       {{ singleDsView?.name || file.name || filename }}
     </mat-card-title>
 
-    <mat-card-subtitle>
+    <mat-card-subtitle class="d-inline-flex align-items-center">
       <mat-icon fontSet="fas" fontIcon="fa-database"></mat-icon>
       <span>
         Dataset preview
       </span>
+      
+      <mat-divider [vertical]="true" class="ml-2 h-2rem"></mat-divider>
+
+      <!-- explore btn -->
+      <a *ngFor="let kgRef of singleDsView.kgReference"
+        [href]="kgRef | doiParserPipe"
+        class="color-inherit"
+        target="_blank">
+        <button mat-icon-button
+          [matTooltip]="singleDsView.EXPLORE_DATASET_IN_KG_ARIA_LABEL">
+          <i class="fas fa-external-link-alt"></i>
+        </button>
+      </a>
+
+      <!-- pin dataset btn -->
+      <ng-container *ngTemplateOutlet="favDatasetBtn; context: { singleDataset: singleDsView }">
+      </ng-container>
+
+      <!-- download zip btn -->
+      <a *ngIf="singleDsView.files && singleDsView.files.length > 0"
+        [href]="singleDsView.dlFromKgHref"
+        class="color-inherit"
+        target="_blank">
+        <button mat-icon-button
+          [matTooltip]="singleDsView.tooltipText"
+          [disabled]="singleDsView.downloadInProgress">
+          <i class="ml-1 fas" [ngClass]="!singleDsView.downloadInProgress? 'fa-download' :'fa-spinner fa-pulse'"></i>
+        </button>
+      </a>
+      
     </mat-card-subtitle>
   </div>
 </mat-card>
@@ -17,11 +47,32 @@
   <single-dataset-view [fullId]="datasetId"
     [hideTitle]="true"
     [hidePreview]="true"
+    [hideExplore]="true"
+    [hidePinBtn]="true"
+    [hideDownloadBtn]="true"
     #singleDsView="singleDatasetView">
 
   </single-dataset-view>
 </mat-card>
 
+<!-- TODO -->
+<!-- this is not exactly right -->
 <div class="mt-4">
   <layer-browser></layer-browser>
 </div>
+
+<!-- templates -->
+<ng-template #favDatasetBtn let-singleDataset="singleDataset">
+  <ng-container *ngTemplateOutlet="isFavCtxTmpl; context: { isFav: (singleDataset.favedDataentries$ | async | datasetIsFaved : singleDataset.dataset) }">
+  </ng-container>
+
+  <ng-template #isFavCtxTmpl let-isFav="isFav">
+    <button mat-icon-button
+      (click)="isFav ? singleDataset.undoableRemoveFav() : singleDataset.undoableAddFav()"
+      [attr.aria-label]="singleDataset.PIN_DATASET_ARIA_LABEL"
+      [matTooltip]="singleDataset.PIN_DATASET_ARIA_LABEL"
+      [color]="isFav ? 'primary' : 'basic'">
+      <i class="fas fa-thumbtack"></i>
+    </button>
+  </ng-template>
+</ng-template>
